+++ lib/mediawiki/butt.rb
@@ -59,8 +59,7 @@ module MediaWiki
       if @logged_in == false
         header['User-Agent'] = 'NotLoggedIn/MediaWiki::Butt'
       else
-        name = get_current_user_name
-        header['User-Agent'] = "#{name}/MediaWiki::Butt"
+        header['User-Agent'] = "#{@name}/MediaWiki::Butt"
       end
 
       res = @client.post(@uri, params, header)